requirements.txt
streamlit
openai
numpy
scikit-learn
python-dotenv